+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - Users can now create, update, and delete Cluster Opaque Ingress Secrets
 - Users can now create and upload to a Schematics Workspace simultaneously via backend route `/api/schematics/newWorkspaceUpload`
 - Added JSON-to-IaC functionality to create Power Edge Router connections for Transit Gateway
+- Users can now connect Power Edge Router enabled workspaces to Transit Gateways
 
 ### Fixes